+++ CHANGELOG.md
@@ -18,6 +18,7 @@ omnibus-gitlab repository.
 - Remove geo_bulk_notify_worker_cron 44def4b5
 - Rework `single_quote` helper as `quote` that can handle escaping
   strings with embedded quotes fdc6a93- Add omniauth sync_profile_from_provider and sync_profile_attributes configuration
+- Add gitlab-ctl pgb-console command
 - Add omniauth sync_profile_from_provider and sync_profile_attributes configuration
 - Upgrade ruby version to 2.3.5
 - Upgrade libyaml version to 0.1.7